Float tempo builds, untested.

+++ sources/Application/Player/SyncMaster.cpp
@@ -20,14 +20,14 @@ void SyncMaster::Start() {
 void SyncMaster::Stop() {
 } ;
 
-void SyncMaster::SetTempo(int tempo) {
+void SyncMaster::SetTempo(float tempo) {
 	tempo_=tempo ;
 	int driverRate=Audio::GetInstance()->GetSampleRate() ;
     playSampleCount_=60.0f*driverRate*2.0f/tempo_/8.0f/float(AUDIO_SLICES_PER_STEP)  ;
 	tickSampleCount_=60.0f*driverRate*2.0f/tempo_/8.0f/float(AUDIO_SLICES_PER_STEP)*tableRatio_  ;
 }  ;
 
-int SyncMaster::GetTempo() {
+float SyncMaster::GetTempo() {
 	return tempo_ ;
 } ;
